fix dependency issue
If a task has a dependency that doesn't exist in
the database yet, we created it, but forgot to fill in the information
when we finally got to loading the real task.

Of course, this is only an issue when loading data from 'task push/task
merge', othwerwise the dependency for a given task will already exist.